Session Aging Changes (#6802)

+++ nic/hal/plugins/cfg/nw/session.hpp
@@ -257,7 +257,7 @@ struct flow_s {
 
 typedef struct flow_state_s {
     uint64_t        create_ts;           // flow create timestamp
-    uint32_t        last_pkt_ts;         // last packet timestamp
+    uint64_t        last_pkt_ts;         // last packet timestamp
     uint64_t        packets;             // packet count on this flow
     uint64_t        bytes;               // byte count on this flow
     uint64_t        drop_packets;        // packets dropped for this flow
